made a useMiniQuiz hook to keep component MiniQuiz clean
