+++ src/pages/miniquiz/NoQuiz.jsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+function NoQuiz() {}
